requirements.txt
aiogram==3.13.1
telethon==1.36.0
opentele==1.15.1
PySocks==1.7.1
SQLAlchemy==2.0.36
alembic==1.14.0
psycopg[binary]==3.2.3
redis==5.1.1
pydantic==2.9.2
pydantic-settings==2.6.1
